BF - Removed decode IOCC from device base; Made BOSC end the current interrupt; Made it possible to check if an interrupt is in the concurrent bag; Tested basic infrastructure processes interrupt resets properly.

+++ 1130.SystemObjects/InterruptManagement/InterruptPool.cs
@@ -22,7 +22,7 @@ namespace S1130.SystemObjects.InterruptManagement
 			return _interruptPool;
 		}
 
-		public Interrupt GetInterrupt(int interruptLevel, IDevice deviceCausingInterrupt, int interruptLevelStatusWord )
+		public Interrupt GetInterrupt(int interruptLevel, IDevice deviceCausingInterrupt, ushort interruptLevelStatusWord )
 		{
 			Interrupt item;
 			if (_interrupts.TryTake(out item)) return item.Setup(interruptLevel, deviceCausingInterrupt, interruptLevelStatusWord);
